[ build ] travis testing
Getting travis to try to build the executable with various
versions of ghc. I have tried to relax some bounds but they
seem to be tight already.

I have reordered the dependencies in herms.cabal by alphabetical
order so that it's easier to maintain.

I also have commented out the part about tests because as far
as I can see there are in fact no tests whatsoever.

The first build takes a bit of time but after that the caching
kicks in and the builds finish in just a couple of minutes.