Reify units to a type
This leads to a consistent internal representation. When dealing with units, we
remove ambiguity at parse-time and never have to deal with synonyms internally.

This has the disadvantage that units aren't presented exactly as the user
entered them (arguably an advantage).

It leads to more compile-time guarantees, when dealing with units GHC can now
make sure our cases are exhaustive.