Fixed formatP parser so that it reads plain string.
The old implementation used [option auto] which relies on Read instance of the returned type.
This doesn't play nicely with String as Read String parses quoted strings like `"yaml"`, not `yaml`.
We can either use [option str] or [strOption] but I chose the former just because it looks nicer to me.