Quantity can now either be a string or number. Fix tests.
Parser shall accept either types and the writer shall output a
number whenever possible.

This change should not be breaking. Old recipe formats shall be
accepted and converted to the new format upon the next write.

The encode -> decode invariance test has been fixed. The issue was
due to not parsing metric units correctly and making "other" units
to lower case. QuickCheck will now attempt to find the simplest
counterexample if it exists for those tests.